# Optimise InitializeTreeViewFromPackage