--- a/GUI/Controls/TreeViewFileSorter.cs
+++ b/GUI/Controls/TreeViewFileSorter.cs
@@ -1,4 +1,5 @@
 using System.Collections;
+using System.Runtime.CompilerServices;
 using System.Windows.Forms;
 
 namespace GUI.Controls
@@ -7,11 +8,14 @@ namespace GUI.Controls
     {
         public int Compare(object x, object y)
         {
-            var tx = x as TreeNode;
-            var ty = y as TreeNode;
+            var tx = Unsafe.As<TreeNode>(x); //Normal c-style casts are actually very slow on something called this often, so we cheat
+            var ty = Unsafe.As<TreeNode>(y);
 
-            var folderx = tx.Tag is TreeViewFolder ? -1 : 1;
-            var foldery = ty.Tag is TreeViewFolder ? -1 : 1;
+            var dataX = Unsafe.As<VrfTreeViewData>(tx.Tag); //Again, perf is king here
+            var dataY = Unsafe.As<VrfTreeViewData>(ty.Tag);
+
+            var folderx = dataX.IsFolder ? -1 : 1;
+            var foldery = dataY.IsFolder ? -1 : 1;
 
             if (folderx != foldery)
             {