… GUI/Controls/TreeViewFolder.cs